Added partial support for peer exchange (PEX)
The client announces its support for the extended message protocol,
and indicates that it wants to receive PEX updates. New peers received
via PEX are added to the list of known peers and treated the same way
as peers obtained from trackers. The support is partial in that no
updates are sent back. IPv6 is not yet implemented.