Changed BitTorrent message processing to a new message system
This client is build using multiple threads, where tasks such as
communicating with a tracker, communicating with a peer, file
operations are all executed in own threads. Currently, inter-thread
communication is a bit of a mess. With this and subsequent commits I
am switching to a more unified approach for inter-thread
communication.

Each thread has a state, which is a CLOS object. Further, each thread
can receive messages, where a message is a list that starts with a
keyword. A single generic function, `call`, is used across all
threads, and it implements dispatch on the thread state and on the CAR
of the message.