Rewrote the scheduler (requesting blocks from peers)
Existing scheduler was rudimentary, and importantly it did not take
the maximum number of block requests that a peer accepts into account.
The new scheduler does that, and further it should recover any lost
blocks.

The new scheduler operates as follows.

Each peer has an associated queue and prequeue. The queue is an
ordered dictionary, which registers blocks requested from the peer.
The prequeue is a doubly linked list, which registers blocks scheduled
for downloading but not yet requested from the peer.

1. A random piece that has not been chosen earlier, that is not yet
verified, and that a peer has, is selected. All blocks from this piece
are added to the prequeue in a random order.

2. All requested blocks that are out of order by more than
*max-delay*, i.e. that have been requested at least *max-delay*
blocks earlier than a block that have been successfully received,
are presumed lost. They are removed from the queue, and added to
the front of the prequeue.

3. Every *clock* cycle, new blocks are drawn from the prequeue,
requested from the peer, and added to the queue so as to maintain the
queue at its maximum capacity. The maximum capacity is initially given
by *default-queue-size*, but it is adjusted if a new capacity is
received via extended BitTorrent handshake. Whenever the prequeue is
empty, it is refilled (see step 1).

4. When requested blocks are received, they are removed from the
queue.

Per design, the prequeues and queues of different peers are not
coordinated. As a result, some blocks will be requested from multiple
peers, creating a small overhead. A piece and its constituent blocks
stop being eligible for requesting only once the piece is verified.